add user.py
