+++ Disteam/steam_api/types/player.py
@@ -1,22 +1,6 @@
 from typing import List, TypedDict
 
 
-class PlayerSummary(TypedDict):
-    steamid: str
-    communityvisibilitystate: int
-    profilestate: int
-    personaname: str
-    lastlogoff: int
-    profileurl: str
-    avatar: str
-    avatarmedium: str
-    avatarfull: str
-
-
-class Players(TypedDict):
-    players: List[PlayerSummary]
-
-
 class RecentGame(TypedDict):
     appid: str
     playtime_forever: int
@@ -29,7 +13,3 @@ class RecentGames(TypedDict):
 
 class PlayerRecentGamesResponse(TypedDict):
     response: RecentGames
-
-
-class PlayerSummaryResponse(TypedDict):
-    response: Players
